Refactor proofGenerator
