+++ src/app/components/Game/Actions/GameActions.jsx
@@ -9,4 +9,3 @@ const GameActions = ({ setResult, confirmProof, resumePause, playing }) => (
 );
 
 export default GameActions;
-
